Extract JsonMapper
